Added sct finished view.

+++ Script_odont/Core/SctFinished.swift
@@ -19,5 +19,5 @@ struct SctFinished
     
     var statistics: SctStatistics
     
-    var score: Int
+    var score: Double
 }
